refactor: remove old direct api
