+++ cmd/cli/commands/deals.go
