+++ cmd/cli/commands/hub_tasks.go
@@ -1,48 +1,21 @@
 package commands
 
 import (
-	"encoding/json"
-	"fmt"
 	"os"
-	"time"
 
-	pb "github.com/sonm-io/core/proto"
 	"github.com/spf13/cobra"
 )
 
 func init() {
-	nodeHubTaskRootCmd.AddCommand(nodeHubTaskListCmd, nodeHubTaskStatusCmd)
+	hubTasksRootCmd.AddCommand(hubTaskListCmd, hubTaskStatusCmd)
 }
 
-var nodeHubTaskRootCmd = &cobra.Command{
+var hubTasksRootCmd = &cobra.Command{
 	Use:   "task",
 	Short: "Operations with tasks",
 }
 
-func printNodeTaskStatus(cmd *cobra.Command, tasksMap map[string]*pb.TaskListReply_TaskInfo) {
-	if isSimpleFormat() {
-		for worker, tasks := range tasksMap {
-			if len(tasks.GetTasks()) == 0 {
-				cmd.Printf("Worker \"%s\" has no tasks\r\n", worker)
-				continue
-			}
-
-			cmd.Printf("Worker \"%s\":\r\n", worker)
-			i := 1
-			for ID, status := range tasks.GetTasks() {
-				up := time.Duration(status.GetUptime())
-				cmd.Printf("  %d) %s \r\n     %s  %s (up: %v)\r\n",
-					i, ID, status.Status.String(), status.ImageName, up.String())
-				i++
-			}
-		}
-	} else {
-		b, _ := json.Marshal(tasksMap)
-		fmt.Printf("%s\r\n", string(b))
-	}
-}
-
-var nodeHubTaskListCmd = &cobra.Command{
+var hubTaskListCmd = &cobra.Command{
 	Use:    "list",
 	Short:  "Show task list",
 	PreRun: loadKeyStoreWrapper,
@@ -63,7 +36,7 @@ var nodeHubTaskListCmd = &cobra.Command{
 	},
 }
 
-var nodeHubTaskStatusCmd = &cobra.Command{
+var hubTaskStatusCmd = &cobra.Command{
 	Use:    "status <task_id>",
 	Short:  "Show task status",
 	Args:   cobra.MinimumNArgs(1),
